geotiff: extract read_geotiff_dask to _backends/dask.py (#1886)
Step 8 of #1813's multi-PR refactor of __init__.py. Pure code motion;
no public API change.

Moved into a new xrspatial/geotiff/_backends/dask.py:

- read_geotiff_dask: dask read entry point (~350 lines), with the
  HTTP/fsspec COG metadata prefetch, MinIsWhite nodata-inversion
  detection, window clipping, max_pixels guard, and 50k-task graph
  cap intact.
- _delayed_read_window: the per-chunk reader closure (~100 lines)
  that fans out via dask.delayed.

Function bodies unchanged except for adjusted relative imports
(``._X`` -> ``.._X``). Two helpers still living in __init__.py
(``read_vrt`` and ``_read_geo_info``) are lazy-imported inside the
function bodies to avoid a circular import; later PRs in #1813 will
move them out and the lazy imports can become static.

Promotes the matching lazy import on the other side: _backends/gpu.py's
``from .. import read_geotiff_dask`` (added for the same circular-
import reason in #1895) becomes a static top-of-module
``from .dask import read_geotiff_dask`` now that the target is at a
sibling path.

Test changes
============

- test_dask_no_op_astype_1624.py monkeypatches ``_read_to_array``; the
  target moves from ``xrspatial.geotiff`` to
  ``xrspatial.geotiff._backends.dask`` so the spy still intercepts the
  per-chunk worker. Same pattern as the GPU spy update in #1895.
- Restored the ``_apply_nodata_mask_gpu`` re-export from __init__.py
  (with a ``# noqa: F401`` marker). PR #1895's review-fix commit
  dropped this re-import after concluding it had no internal callers;
  it overlooked five test imports across ``test_nodata_nan_int_1774``
  and ``test_nodata_out_of_range_1581`` that exercise the helper
  directly. Restoring the re-export fixes 6 tests that were failing
  on main since #1895 merged.

Diff stats
==========

- __init__.py: 2962 -> 2504 (-458 lines), plus 1 re-import line.
- New _backends/dask.py: 481 lines.

Verification: pixel-parity matrix from #1889 (192 cells), runtime
identity (9), dask int nodata chunks (#1597), attrs parity (#1548),
dask no-op astype (#1624), gpu chunks out-of-core (#1876), full
geotiff suite minus the 8 pre-existing main failures -- 2862/2862 pass.

Refs #1813.
